+++ src/hunter/exceptions.py
@@ -1,0 +1,2 @@
+class BadPredicateComposition(Exception):
+    pass